+++ CHANGELOG.md
@@ -6,3 +6,9 @@
 * I2C1 on FPGA Machine
 * I2C1 multi-master implementation with the CPU
 * FAN Control
+
+## 1.1 (0xdeb0004)
+* Fixed TWI byte swapping (R/T)
+* Added i2c1-i2c2-i2c3 devices in FPGA shadow register
+* Change iMX6 TWI bus master handler
+* Change TWI execution timing